requirements.txt
python >= 3.6.8
pytorch >= 1.0.0
numpy >= 1.15.4